# src/openapi/post.v1.yaml
openapi: 3.0.0
info:
  title: 投稿管理API
  description: 投稿管理API
  version: '1.0'
  contact:
    name: yuta-katayama-23
    email: yuta-katayama-23@example.com
servers:
  - url: 'http://localhost:3000'
    description: local development
tags:
  - name: Post API
paths:
  /api/v1/post:
    post:
      summary: 新規投稿作成API
      operationId: get-api-v1-post
      description: 新規に投稿を作成します。
      tags:
        - Post API
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: タイトル
                  maxLength: 64
                  minLength: 1
              required:
                - title
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int32
                    minimum: 1
                    example: 1
                    description: 投稿ID
                  title:
                    type: string
                    description: タイトル
                    minLength: 1
                    maxLength: 64
                  like_count:
                    type: integer
                    format: int32
                    minimum: 0
                    example: 1
                    description: いいねの数
                  comment_count:
                    type: integer
                    format: int32
                    minimum: 0
                    example: 1
                    description: コメントの数
                  updated_at:
                    type: integer
                    description: 更新日時（UNIX TIME：秒）
                    example: 1639444400
                    minimum: 1600000000
                    format: int64
                  created_at:
                    type: integer
                    format: int64
                    description: 作成日時（UNIX TIME：秒）
                    example: 1639444400
                    minimum: 1600000000
                required:
                  - id
                  - title
                  - like_count
                  - comment_count
                  - updated_at
                  - created_at
        '400':
          $ref: '#/components/responses/error'
        '401':
          $ref: '#/components/responses/error'
        '403':
          $ref: '#/components/responses/error'
        '404':
          $ref: '#/components/responses/error'
        '409':
          $ref: '#/components/responses/error'
        '415':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  '/api/v1/post/{post_id}/comment':
    parameters:
      - schema:
          type: integer
          format: int32
          minimum: 1
          example: 1
        name: post_id
        in: path
        required: true
        description: 投稿ID
    post:
      summary: 投稿へのコメント作成API
      description: 投稿に対してコメントします。
      operationId: post-api-v1-post-comment
      tags:
        - Post API
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: コメントの内容
                  minLength: 1
                  maxLength: 256
              required:
                - content
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                    format: int32
                    minimum: 1
                    description: コメントID
                  post_id:
                    type: integer
                    format: int32
                    example: 1
                    minimum: 1
                    description: 投稿ID
                  content:
                    type: string
                    description: コメントの内容
                    minLength: 1
                    maxLength: 256
                  updated_at:
                    type: integer
                    description: 更新日時（UNIX TIME：秒）
                    example: 1639444400
                    minimum: 1600000000
                    format: int64
                  created_at:
                    type: integer
                    format: int64
                    description: 作成日時（UNIX TIME：秒）
                    example: 1639444400
                    minimum: 1600000000
                required:
                  - id
                  - post_id
                  - content
                  - updated_at
                  - created_at
        '400':
          $ref: '#/components/responses/error'
        '401':
          $ref: '#/components/responses/error'
        '403':
          $ref: '#/components/responses/error'
        '404':
          $ref: '#/components/responses/error'
        '409':
          $ref: '#/components/responses/error'
        '415':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
components:
  schemas: {}
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: ログイン時に取得されるトークン
  responses:
    error:
      description: ''
      content:
        application/json:
          schema:
            $ref: ./error.v1.yaml#/components/schemas/error